+++ base/flapping.cc
@@ -20,15 +20,14 @@
 
 /*********** COMMON HEADER FILES ***********/
 
-#include "configuration.hh"
-
-#include "config.hh"
-#include "common.hh"
-#include "objects.hh"
+#include "conf.hh"
 #include "comments.hh"
 #include "statusdata.hh"
 #include "nagios.hh"
 #include "broker.hh"
+#include "notifications.hh"
+#include "configuration.hh"
+#include "flapping.hh"
 
 extern com::centreon::scheduler::configuration config;
 
@@ -48,8 +47,8 @@ extern unsigned long    modified_service_process_attributes;
 void check_for_service_flapping(service *svc, int update, int allow_flapstart_notification){
 	int update_history=TRUE;
 	int is_flapping=FALSE;
-	register int x=0;
-	register int y=0;
+	register unsigned int x=0;
+	register unsigned int y=0;
 	int last_state_history_value=STATE_OK;
 	double curved_changes=0.0;
 	double curved_percent_change=0.0;
@@ -173,8 +172,8 @@ void check_for_service_flapping(service *svc, int update, int allow_flapstart_no
 void check_for_host_flapping(host *hst, int update, int actual_check, int allow_flapstart_notification){
 	int update_history=TRUE;
 	int is_flapping=FALSE;
-	register int x=0;
-	register int y=0;
+	register unsigned int x=0;
+	register unsigned int y=0;
 	int last_state_history_value=HOST_UP;
 	unsigned long wait_threshold=0L;
 	double curved_changes=0.0;
@@ -215,7 +214,7 @@ void check_for_host_flapping(host *hst, int update, int actual_check, int allow_
 	        }
 
 	/* if we didn't have an actual check, only update if we've waited long enough */
-	if(update_history==TRUE && actual_check==FALSE && (current_time-hst->last_state_history_update)<wait_threshold){
+	if(update_history==TRUE && actual_check==FALSE && static_cast<unsigned long>(current_time-hst->last_state_history_update)<wait_threshold){
 
 		update_history=FALSE;
